+++ CHANGELOG.md
@@ -8,8 +8,9 @@
 
 ### Enhancements:
 
-- feat(logging): Add support for 'CompressionCodec' and 'GzipLevel' attribute to the HTTPS endpoint.
+- feat(logging): Add support for 'CompressionCodec' and 'GzipLevel' attribute to the HTTPS endpoint.[#734](https://github.com/fastly/go-fastly/pull/734)
 - feat(domainmanagement): Adds support for domain-management/v1/domains endpoints [#726](https://github.com/fastly/go-fastly/pull/726)
+- feat(ngwaf/rules): adds support for the deception action type [#736](https://github.com/fastly/go-fastly/pull/736)
 
 ### Bug fixes: