+++ CHANGELOG.md
@@ -7,13 +7,15 @@
 ### Breaking:
 
 ### Enhancements:
-feat(logging): Add support for 'CompressionCodec' and 'GzipLevel' attribute to the HTTPS endpoint.
 
+- feat(logging): Add support for 'CompressionCodec' and 'GzipLevel' attribute to the HTTPS endpoint.[#734](https://github.com/fastly/go-fastly/pull/734)
 - feat(domainmanagement): Adds support for domain-management/v1/domains endpoints [#726](https://github.com/fastly/go-fastly/pull/726)
+- feat(ngwaf/rules): adds support for the deception action type and the allow_interactive field [#736](https://github.com/fastly/go-fastly/pull/736)
 
 ### Bug fixes:
 
 ### Dependencies:
+
 - build(deps): `actions/checkout` from 4 to 5 ([#735](https://github.com/fastly/go-fastly/pull/735))
 
 ## [v11.1.2](https://github.com/fastly/go-fastly/releases/tag/v11.1.2) (2025-08-06)